requirements.txt
# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2
xgboost==1.7.6
lightgbm==4.0.0
statsmodels==0.14.0

# Geospatial analysis
geopandas==0.13.2
folium==0.14.0
pyproj==3.6.0
rasterio==1.3.8
xarray==2023.7.0
rioxarray==0.14.1

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
dash==2.13.0
bokeh==3.2.2

# Machine learning and statistics
tensorflow==2.13.0
pytorch==2.0.1
prophet==1.1.4
tslearn==0.6.2
pmdarima==2.0.3

# Web application
flask==2.3.3
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
flask-login==0.6.2
dash-bootstrap-components==1.4.2

# Database
sqlalchemy==2.0.20
pymysql==1.1.0
psycopg2-binary==2.9.7

# Utilities
joblib==1.3.2
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
jupyterlab==4.0.5
nbformat==5.9.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0